requirements.txt
django
django-rest-framework
web3
py-solc